Refactor chain:forks to use forkCounter
